Refactor template.module.scss
