+++ components/Footer/Footer.tsx
@@ -1,28 +1,28 @@
-import Grid from "../styles/modules/grid.module.scss";
-import Template from "../styles/modules/template.module.scss";
+import Grid from "../../styles/modules/grid.module.scss";
+import Style from "./styles.module.scss";
 import Image from "next/image";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import Link from "next/link";
-import logo from "../img/logo.svg";
-import forbes from "../img/forbes.png";
+import logo from "../../img/logo.svg";
+import forbes from "../../img/forbes.png";
 import { faLinkedinIn, faTwitter } from "@fortawesome/free-brands-svg-icons";
 import React from "react";
-import { pageSettings as termsOfServicePageSettings } from "../pages/termsofservice";
-import { pageSettings as privacyPolicyPageSettings } from "../pages/privacypolicy";
-import { pageSettings as legalPageSettings } from "../pages/legal";
-import { pageSettings as careersPageSettings } from "../pages/careers";
-import { pageSettings as newsroomPageSettings } from "../pages/newsroom";
+import { pageSettings as termsOfServicePageSettings } from "../../pages/termsofservice";
+import { pageSettings as privacyPolicyPageSettings } from "../../pages/privacypolicy";
+import { pageSettings as legalPageSettings } from "../../pages/legal";
+import { pageSettings as careersPageSettings } from "../../pages/careers";
+import { pageSettings as newsroomPageSettings } from "../../pages/newsroom";
 import clsx from "clsx";
 
 const Footer: React.FunctionComponent = () => {
   return (
-    <footer id="contact" className={clsx(Template.footer, "section")}>
+    <footer id="contact" className={clsx(Style.footer, "section")}>
       <div
         className={clsx(
           Grid.marginCenter,
           Grid.container,
           Grid.row,
-          Template.footerTitle
+          Style.footerTitle
         )}
       >
         <h3>Get in Touch With Us.</h3>
@@ -33,18 +33,18 @@ const Footer: React.FunctionComponent = () => {
           Grid.container,
           Grid.row,
           Grid.gap,
-          Template.footerContent
+          Style.footerContent
         )}
       >
-        <div className={clsx(Template.footerFormContainer, "flow")}>
-          <div className={clsx(Template.footerText)}>
+        <div className={clsx(Style.footerFormContainer, "flow")}>
+          <div className={clsx(Style.footerText)}>
             <p>Find out how Nexus can unlock your productivity.</p>
           </div>
 
           <form
             action="https://formspree.io/f/mvolablv"
             method="POST"
-            className={clsx(Template.footerForm, "flow")}
+            className={clsx(Style.footerForm, "flow")}
           >
             <div className={clsx(Grid.rowWrapoff)}>
               <input
@@ -76,16 +76,16 @@ const Footer: React.FunctionComponent = () => {
             </div>
             <div
               className={clsx(
-                Template.footerFormSubmit,
+                Style.footerFormSubmit,
                 Grid.row,
                 Grid.gap,
                 "flow"
               )}
             >
-              <button className={clsx(Template.footerFormButton)} type="submit">
+              <button className={clsx(Style.footerFormButton)} type="submit">
                 Send
               </button>
-              <div className={clsx(Template.socialIcons, Grid.row, "")}>
+              <div className={clsx(Style.socialIcons, Grid.row, "")}>
                 <a
                   href="https://www.linkedin.com/company/torchai"
                   target="_blank"
@@ -105,7 +105,7 @@ const Footer: React.FunctionComponent = () => {
           </form>
         </div>
 
-        <div className={clsx(Template.footerTerms)}>
+        <div className={clsx(Style.footerTerms)}>
           <ul>
             <li>
               <Link href={newsroomPageSettings.path}>
@@ -135,7 +135,7 @@ const Footer: React.FunctionComponent = () => {
           </ul>
         </div>
 
-        <div className={clsx(Template.footerAddress)}>
+        <div className={clsx(Style.footerAddress)}>
           <ul>
             <li>5250 W 116TH Pl Suite 200</li>
             <li>Leawood, KS, 66211</li>
@@ -156,7 +156,7 @@ const Footer: React.FunctionComponent = () => {
 
       <div
         className={clsx(
-          Template.footerCopy,
+          Style.footerCopy,
           Grid.marginCenter,
           Grid.container,
           Grid.betweenXs,
@@ -164,7 +164,7 @@ const Footer: React.FunctionComponent = () => {
         )}
       >
         <Image
-          className={clsx(Template.logo)}
+          className={clsx(Style.logo)}
           src={logo}
           alt="The logo of Torch.ai"
         />
