+++ pages/careers/index.tsx
@@ -2,7 +2,7 @@
 
 import React, { ReactElement, useEffect } from "react";
 import Image from "next/image";
-import Footer from "../../components/Footer";
+import Footer from "../../components/Footer/Footer";
 import Head from "next/head";
 import forbes from "../../img/forbes.png";
 import PageHeader from "../../components/PageHeader/PageHeader";
