+++ pages/newsroom/index.tsx
@@ -4,7 +4,7 @@ import { TypeNews } from "../../generated/contentful";
 import { getAllNewsEntries } from "../../contentful/client";
 import { getHeadPageTitle } from "../../utils/meta";
 import Head from "next/head";
-import Footer from "../../components/Footer";
+import Footer from "../../components/Footer/Footer";
 import NewsGrid from "../../components/News/NewsGrid";
 import Style from "./styles.module.scss";
 import LayoutContext from "../../components/layout/LayoutContext";
